third_party/blink/web_tests/inspector-protocol/css/css-set-effective-property-value-expected.txt

The test verifies functionality of protocol method CSS.setEffectivePropertyValueForNode.
Dumping inline style: 
{
    padding-top: 55px;
    margin-top: 33px !important;
    --x: foo;
    padding-top: 55px;
    --x: foo;
    margin-top: 33px !important;
}
Dumping matched rules: 
*#inspected* {    regular
    margin-left: 10px !important;
    margin-left: 10px !important;
}
*#inspected* {    regular
    padding: 10px 20px 30px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 30px;
    padding-left: 40px;
    padding-top: 50px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
    }
Dumping inherited rules: 

Running test: testBasicPropertyChange
{
    padding-top: 55px;
    margin-top: 33px !important;
    --x: foo;
    padding-top: 55px;
    --x: foo;
    margin-top: 33px !important;
}
*#inspected* {    regular
    margin-left: 10px !important;
    margin-left: 10px !important;
}
*#inspected* {    regular
    padding: 10px 20px 30px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 30px;
    padding-left: 40px;
    padding-top: 50px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 101px;
        margin-top: 15px !important;
        margin-left: 20px;
        padding-left: 101px;
        margin-top: 15px !important;
    }

Running test: testChangePropertyInShortHand
{
    padding-top: 55px;
    margin-top: 33px !important;
    --x: foo;
    padding-top: 55px;
    --x: foo;
    margin-top: 33px !important;
}
*#inspected* {    regular
    margin-left: 10px !important;
    margin-left: 10px !important;
}
*#inspected* {    regular
    padding: 10px 20px 101px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 101px;
    padding-left: 40px;
    padding-top: 50px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
    }

Running test: testChangeImportantProperty
{
    padding-top: 55px;
    margin-top: 33px !important;
    --x: foo;
    padding-top: 55px;
    --x: foo;
    margin-top: 33px !important;
}
*#inspected* {    regular
    margin-left: 101px !important;
    margin-left: 101px !important;
}
*#inspected* {    regular
    padding: 10px 20px 30px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 30px;
    padding-left: 40px;
    padding-top: 50px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
    }

Running test: testChangeInlineProperty
{
    padding-top: 101px;
    margin-top: 33px !important;
    --x: foo;
    padding-top: 101px;
    --x: foo;
    margin-top: 33px !important;
}
*#inspected* {    regular
    margin-left: 10px !important;
    margin-left: 10px !important;
}
*#inspected* {    regular
    padding: 10px 20px 30px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 30px;
    padding-left: 40px;
    padding-top: 50px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
    }

Running test: testChangeInlineImportantProperty
{
    padding-top: 55px;
    margin-top: 101px !important;
    --x: foo;
    padding-top: 55px;
    --x: foo;
    margin-top: 101px !important;
}
*#inspected* {    regular
    margin-left: 10px !important;
    margin-left: 10px !important;
}
*#inspected* {    regular
    padding: 10px 20px 30px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 30px;
    padding-left: 40px;
    padding-top: 50px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
    }

Running test: testChangeMissingProperty
{
    padding-top: 55px;
    margin-top: 33px !important;
    --x: foo;
    margin-bottom: 101px;
    padding-top: 55px;
    --x: foo;
    margin-bottom: 101px;
    margin-top: 33px !important;
}
*#inspected* {    regular
    margin-left: 10px !important;
    margin-left: 10px !important;
}
*#inspected* {    regular
    padding: 10px 20px 30px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 30px;
    padding-left: 40px;
    padding-top: 50px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
    }

Running test: testAppendWithSeparator
Dumping inline style: 
{
    padding-left: 10px;
    padding-left: 10px;
}
Dumping matched rules: 
Dumping inherited rules: 
Resulting styles
{
    padding-left: 10px;
    padding-right: 101px;
    padding-left: 10px;
    padding-right: 101px;
}

Running test: testChangeCustomProperty
{
    padding-top: 55px;
    margin-top: 33px !important;
    --x: bar;
    padding-top: 55px;
    --x:  bar;
    margin-top: 33px !important;
}
*#inspected* {    regular
    margin-left: 10px !important;
    margin-left: 10px !important;
}
*#inspected* {    regular
    padding: 10px 20px 30px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 30px;
    padding-left: 40px;
    padding-top: 50px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
        margin-left: 20px;
        padding-left: 10px;
        margin-top: 15px !important;
    }
